# secp256k1: Reduce scalar base mult copies.