~**This is rebased on #2888**~.

Profiling shows that around 7.5% of the time in scalar base multiplication is attributed to `duffcopy`.  Upon further examination, this is the result of a combination of the range statement making copies of the bytes and the need to construct a Jacobian point from the individual field values stored in the in-memory byte points table.

This optimizes the function to avoid that as follows:

- Perform the conversion to Jacobian once when the affine byte table is decompressed from the stored values
- Make use of those Jacobian points directly
- Use an indexed for loop instead of a range over the bytes 
- Perform the calculation using the result variable directly instead of via a local variable that is copied to the result

The following benchmark results show the speedup is in line with the expected gains per the profiling results:

```
name                     old time/op   new time/op    delta
------------------------------------------------------------------------------
ScalarBaseMultNonConst   24.1µs ±22%   22.5µs ± 2%   -6.97%  (p=0.000 n=98+96)
```